move fixtures/api/graphql
