--- test/fixtures/api/graphql/commit-empty-nodes.json
@@ -1,13 +1,0 @@
-{
-  "data": {
-    "repository": {
-      "ref": {
-        "target": {
-          "history": {
-            "edges": []
-          }
-        }
-      }
-    }
-  }
-}
