--- test/fixtures/api/graphql/pull-request-empty-nodes.json
@@ -1,9 +1,0 @@
-{
-  "data": {
-    "repository": {
-      "pullRequests": {
-        "edges": []
-      }
-    }
-  }
-}
